added BaseFragment.kt
