+++ company/src/main/java/com/amk/company/ui/CompanyFragment.kt
@@ -1,34 +1,24 @@
 package com.amk.company.ui
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.widget.Toast
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.amk.company.R
 import com.amk.company.databinding.FragmentCompanyBinding
 import com.amk.company.ui.candlechart.CandlestickViewImpl
 import com.amk.company.viewmodel.CompanyViewModel
 import com.amk.core.repository.Repository
+import com.amk.core.ui.BaseFragment
 import org.koin.android.ext.android.inject
 
-class CompanyFragment : Fragment() {
-    private var _binding: FragmentCompanyBinding? = null
-    private val binding get() = _binding!!
+class CompanyFragment : BaseFragment<FragmentCompanyBinding>() {
+    override fun getViewBinding() = FragmentCompanyBinding.inflate(layoutInflater)
+
     private val repository: Repository by inject()
     private lateinit var viewModel: CompanyViewModel
     private val candlestickView: CandlestickViewImpl by lazy {
         layoutInflater.inflate(R.layout.view_candlestick, null) as CandlestickViewImpl
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentCompanyBinding.inflate(inflater, container, false)
-        return binding.root
     }
 
     override fun onPause() {
