+++ core/build.gradle
@@ -29,6 +29,9 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
+    buildFeatures {
+        viewBinding true
+    }
 }
 
 dependencies {
@@ -62,5 +65,6 @@ dependencies {
     //Room
     implementation 'androidx.room:room-common:2.5.0-alpha02'
     implementation 'androidx.room:room-ktx:2.5.0-alpha02'
+    implementation 'androidx.room:room-runtime:2.5.0-alpha02'
     kapt 'androidx.room:room-compiler:2.5.0-alpha02'
 }
