+++ core/src/main/java/com/amk/core/ui/BaseFragment.kt
@@ -1,0 +1,28 @@
+package com.amk.core.ui
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.viewbinding.ViewBinding
+
+abstract class BaseFragment<VB : ViewBinding> : Fragment() {
+    private var _binding: VB? = null
+    val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = getViewBinding()
+        return binding.root
+    }
+
+    abstract fun getViewBinding(): VB
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
