+++ list-companies/src/main/java/com/amk/mylibrary/presentation/adapter/ListCompaniesAdapter.kt
@@ -3,7 +3,6 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.amk.core.entity.Company
-import com.amk.core.entity.EntityCompany
 import com.amk.mylibrary.databinding.ItemCompanyBinding
 
 
@@ -14,7 +13,7 @@ class ListCompaniesAdapter(
     RecyclerView.Adapter<ListCompaniesHolder>() {
 
     interface OnStateClickListener {
-        fun onStateClick(entityCompany: Company, position: Int)
+        fun onStateClick(company: Company, position: Int)
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ListCompaniesHolder {
@@ -23,10 +22,10 @@ class ListCompaniesAdapter(
     }
 
     override fun onBindViewHolder(holder: ListCompaniesHolder, position: Int) {
-        val entityCompany: Company = list[position]
-        holder.bind(entityCompany)
+        val company: Company = list[position]
+        holder.bind(company)
         holder.itemView.setOnClickListener {
-            onClickListener.onStateClick(entityCompany, position)
+            onClickListener.onStateClick(company, position)
         }
     }
