+++ list-companies/src/main/java/com/amk/mylibrary/ui/ListCompanyFragment.kt
@@ -1,13 +1,10 @@
 package com.amk.mylibrary.ui
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.animation.Animation
 import android.view.animation.AnimationUtils
 import android.widget.Toast
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
@@ -20,6 +17,7 @@ import com.amk.core.repository.NetworkRepository
 import com.amk.core.repository.RepositoryCompany
 import com.amk.core.repository.RepositoryCompanyImpl
 import com.amk.core.retrofit.MoexApiImpl
+import com.amk.core.ui.BaseFragment
 import com.amk.mylibrary.R
 import com.amk.mylibrary.databinding.FragmentListCompanyBinding
 import com.amk.mylibrary.presentation.adapter.ListCompaniesAdapter
@@ -27,30 +25,20 @@ import com.amk.mylibrary.viewmodel.CompaniesListViewModel
 import com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
 import org.koin.android.ext.android.inject
 
-class ListCompanyFragment : Fragment() {
-    private var _binding: FragmentListCompanyBinding? = null
-    private val binding get() = _binding!!
+class ListCompanyFragment : BaseFragment<FragmentListCompanyBinding>() {
+    override fun getViewBinding() = FragmentListCompanyBinding.inflate(layoutInflater)
 
     private lateinit var repository: RepositoryCompany
     private lateinit var viewModel: CompaniesListViewModel
     private var companiesList = mutableListOf<Company>()
     private val coordinator: AppNavigation by inject()
 
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentListCompanyBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         val networkRepository = NetworkRepository(MoexApiImpl().getMoexService())
         val database = DataBaseCacheCompany.getDatabase(requireContext())
         val cacheRepository = CacheRepository(database.cacheDao())
-        repository = RepositoryCompanyImpl(requireContext(),networkRepository, cacheRepository)
+        repository = RepositoryCompanyImpl(requireContext(), networkRepository, cacheRepository)
         viewModel = ViewModelProvider(requireActivity())[CompaniesListViewModel::class.java]
         viewModel.setRepo(repository)
         viewModel.companiesListDataYesterday.observe(viewLifecycleOwner) {
@@ -66,11 +54,6 @@ class ListCompanyFragment : Fragment() {
 
             dialog.show(childFragmentManager, "ok")
         }
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
     }
 
     private fun setRecyclerView(list: List<Company>) {
@@ -98,6 +81,7 @@ class ListCompanyFragment : Fragment() {
             }
         })
     }
+
     private fun hide(fab: ExtendedFloatingActionButton) {
         fab.startAnimation(toBottomAnimation)
         fab.startAnimation(toBottomAnimation)
